# lib/api-zod/src/generated/types/noteUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Traveloop – Personalized Travel Planning API
 * OpenAPI spec version: 0.1.0
 */

export interface NoteUpdate {
  /** @minLength 1 */
  content?: string;
}